+++ HISTORY.md
@@ -7,7 +7,8 @@
 * Morphic: added option to include alpha in color comparison
 * Objects: fixed "set pen color (color picker)" to observe, i.e. overwrite transparency
 * Lists, Threads, Objects: added (Bernat's) JSON parser to SPLIT block
-* Lists, Objects: added "export as JSON" capability 
+* Lists, Objects: added "export as JSON" capability
+* Lists, Objects: automatically parse json files on import 
 
 ### 2019-01-08
 * Objects: automatically parse csv files on import, experimental "raw data" and "parse" ops